+++ core/graphicsManager/graphicsLinker.cpp
@@ -95,7 +95,7 @@ void graphicsLinker::createFramebuffers(){
             framebufferInfo.sType = VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO;
             framebufferInfo.renderPass = renderPass;
             framebufferInfo.attachmentCount = 1;
-            framebufferInfo.pAttachments = &swapChainKHR->attachment(Image).imageView;
+            framebufferInfo.pAttachments = &swapChainKHR->attachment(static_cast<uint32_t>(Image)).imageView;
             framebufferInfo.width = imageExtent.width;
             framebufferInfo.height = imageExtent.height;
             framebufferInfo.layers = 1;
